Change the type for the time variables (#146)
* perf: change the time variables type to "uint64"

refactor: add type conversions in the "getWithdrawableAmount" function
test: add "assertEq" function for the "uint64" type
test: add "createDynamicArray" functions for the "uint64" type
test: change the time variables type to "uint64"

* refactor: declare variables only when needed in scope

chore: visually group struct members with ascii box-drawing chars
refactor: define variables alphabetically
refactor: make "currentTime" of type "uint64"
test: add "LogNamedArray" event for "uint64[]"

+++ test/unit/abstract-sablier-v2/AbstractSablierV2UnitTest.t.sol
@@ -46,12 +46,12 @@ contract AbstractSablierV2 is SablierV2 {
         return address(0);
     }
 
-    function getStartTime(uint256 streamId) external pure override returns (uint256) {
+    function getStartTime(uint256 streamId) external pure override returns (uint64) {
         streamId;
         return 0;
     }
 
-    function getStopTime(uint256 streamId) external pure override returns (uint256) {
+    function getStopTime(uint256 streamId) external pure override returns (uint64) {
         streamId;
         return 0;
     }
